contain and continue color codes
If a column contains color codes, we should keep the color contained
to the column itself, stopping it at the column boundary and starting
it again on the next continuation.

This may have a more subtle bug, related to padding between end of
data in a column and its boundary, when the color codes have added
a background color.